+++ pymicro_wakeword/__init__.py
@@ -1,8 +1,10 @@
 """Tensorflow-based wake word detection."""
+
 from .const import Model
-from .microwakeword import MicroWakeWord
+from .microwakeword import MicroWakeWord, MicroWakeWordFeatures
 
 __all__ = [
     "MicroWakeWord",
+    "MicroWakeWordFeatures",
     "Model",
 ]