+++ tests/test_microwakeword.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from pymicro_wakeword import MicroWakeWord, Model
+from pymicro_wakeword import MicroWakeWord, Model, MicroWakeWordFeatures
 
 _DIR = Path(__file__).parent
 
@@ -26,49 +26,34 @@ def _load_wav(model_name: str, number: int) -> bytes:
 @pytest.mark.parametrize(
     "model,number", list(itertools.product(_MODELS, range(1, _NUM_WAVS + 1)))
 )
-def test_process_clip(model: Model, number: int) -> None:
-    """Test full clip processing."""
-    mww = MicroWakeWord.from_builtin(model)
-
-    # positive
-    audio_bytes = _load_wav(model.value, number)
-    assert mww.process_clip(audio_bytes).detected, (model.value, number)
-
-    # Use other wake word samples as negative samples
-    for other_model in _MODELS:
-        if model == other_model:
-            continue
-
-        audio_bytes = _load_wav(other_model.value, number)
-        assert not mww.process_clip(audio_bytes).detected, (
-            model.value,
-            other_model.value,
-            number,
-        )
-
-
-@pytest.mark.parametrize(
-    "model,number", list(itertools.product(_MODELS, range(1, _NUM_WAVS + 1)))
-)
 def test_process_streaming(model: Model, number: int) -> None:
     """Test streaming processing."""
     mww = MicroWakeWord.from_builtin(model)
+    mww_features = MicroWakeWordFeatures()
 
     # positive
     audio_bytes = _load_wav(model.value, number)
-    assert mww.process_streaming(audio_bytes), (model.value, number)
+    detected = False
+    for features in mww_features.process_streaming(audio_bytes):
+        if mww.process_streaming(features):
+            detected = True
+            break
+
+    assert detected, (model.value, number)
 
     # Use other wake word samples as negative samples
     for other_model in _MODELS:
         if model == other_model:
             continue
 
-        # Reset between uses
+        detected = False
         mww.reset()
+        mww_features.reset()
 
         audio_bytes = _load_wav(other_model.value, number)
-        assert not mww.process_streaming(audio_bytes), (
-            model.value,
-            other_model.value,
-            number,
-        )
+        for features in mww_features.process_streaming(audio_bytes):
+            if mww.process_streaming(features):
+                detected = True
+                break
+
+        assert not detected, (model.value, other_model.value, number)