+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.0
+
+- Use embedded TFLite library with ctypes
+- Separate out feature processing
+- Remove `process_clip`
+- Include Alexa model
+- Transition to pyproject.toml
+
 ## 1.0.0
 
 - Initial release